Pick a year-specific table from among lookup.acs2022, etc.

Pick a year-specific lookup table of information on American Community Survey (ACS) tables, from the Census Bureau, namely which sequence files on the FTP site contain which tables and which variables.

Usage

get.lookup.acs(end.year = acsdefaultendyearhere_func())

Arguments

end.year

Character, optional, like "2020", which specifies the 2016-2020 dataset. Defines which 5-year summary file to use, based on end-year. Note: Function stops with error if given end.year is not yet added to this package or is too old.

Value

returns a data.frame

See also

Examples

 names(lookup.acs2022)
 names(lookup.acs)