Get name of Census file with Sequence and Table Numbers for ACS 5-year summary file data

Returns name of text file provided by US Census Bureau, such as Sequence_Number_and_Table_Number_Lookup.txt, which provides the sequence numbers (file numbers) and table numbers for data in the American Community Survey (ACS) 5-year summary file.

Usage

get.lookup.file.name(end.year = acsdefaultendyearhere_func())

Arguments

end.year

Not yet implemented, but will be optional end year for 5-year summary file, as character

Value

Returns character element that is name of file such as "Sequence_Number_and_Table_Number_Lookup.txt"

See also