Obtain American Community Survey Summary File Data Tables from Census Bureau FTP Site

THIS PACKAGE MAY BE MOSTLY OBSOLETE NOW THAT SOME CRAN PACKAGES OFFER MOST OF THESE TOOLS.

Details

There is a package called totalcensus that tries to offer tools to download ACS5yr tables, but the CRAN and dev github versions have bugs as of 8/2023, and did not work out of the box. totalcensus https://cran.r-project.org/web/packages/totalcensus/index.html see devtools::install_github("GL-Li/totalcensus")

ACSdownload is an R package that helps you download and parse huge raw data from the Census Bureau American Community Survey 5-year Summary Files, providing demographic data at the block and tract levels of resolution.

You can obtain data from the entire USA all at once using this package, for one or more tables.

Typically the Census Bureau makes it easy to obtain data from one state at a time, not every block group in the US. There are roughly 220,000 block groups in the US, and around 74,000 tracts

Some options for obtaining Census ACS data are listed here: http://www.census.gov/programs-surveys/acs/data.html Other data sources that may be relevant include Census geodatabases at http://www.census.gov/geo/maps-data/data/tiger-data.html and data at http://www.census.gov/geo/maps-data/data/gazetteer.html For any imported/suggested packages not on CRAN, see https://ejanalysis.github.io

Key function is get.acs(), and also see set.needed(), nhgis(), and also see data(lookup.acs2020) HOWEVER, DATA FORMAT IS CHANGING FOR SUMMARY FILE ACS DATA: https://www.census.gov/programs-surveys/acs/data/summary-file/ updates-to-acs-summary-file.Overview.html

Typically the Census Bureau makes it easy to obtain data from one state at a time, not every block group in the US. There are roughly 220,000 block groups in the US, and around 74,000 tracts. The key function in this package is get.acs()

For ACS documentation, see http://www.census.gov/programs-surveys/acs.html

Several options for obtaining Census ACS data are now listed here:

http://www.census.gov/programs-surveys/acs/data.html
Limits on downloads via American Fact Finder, not all US tracts at once are noted here: https://ask.census.gov/faq.php?id=5000&faqId=1653

Other data sources that may be relevant include Census geodatabases at
http://www.census.gov/geo/maps-data/data/tiger-data.html and data at
http://www.census.gov/geo/maps-data/data/gazetteer.html.

Also see the help for [get.acs()] \cr\cr

References

http://ejanalysis.github.io
http://www.ejanalysis.com

See also

Author

info@ejanalysis.com