ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset

This data set provides information about variables in tables from the latest American Community Survey (ACS) 5-year summary file.

Format

A data.frame with these fields:

data.frame: > 25,000 obs. of approx 9 or 10 variables:

- $ File.ID               : chr  "ACSSF"  ...
- $ File.ID                : chr  "ACSSF" "ACSSF" "ACSSF" "ACSSF" ...
- $ Table.ID               : chr  "B01001" "B01001" "B01001" "B01001" ...
- $ Sequence.Number        : chr  "0001" "0001" "0001" "0001" ...
- $ Line.Number            : num  NA NA 1 2 3 4 5 6 7 8 ...
- $ Start.Position         : num  7 NA NA NA NA NA NA NA NA NA ...
- $ Total.Cells.in.Table   : chr  "49 CELLS" NA NA NA ...
- $ Total.Cells.in.Sequence: num  NA NA NA NA NA NA NA NA NA NA ...
- $ Table.Title            : chr  "SEX BY AGE" "Universe:  Total population" "Total:" "Male:" ...
- $ Subject.Area           : chr  "Age-Sex" NA NA NA ...

Details

See attributes(lookup.acs) to check vintage of dataset.

The data and documentation for the 5 years ending in year X is typically available by December of the year X+1.

Created using a script in the source package:

ACSdownload/data-raw/datacreate_lookup.acs.R

See also